src/data/real_data_collector.py
"""
Real Data Collector for Weather-Driven Electricity Price Forecasting
Modules for Meteostat, IEX, and Grid-India data ingestion
"""

import os
import sys
import pandas as pd
import numpy as np
from datetime import datetime, timedelta
from meteostat import hourly, Point, stations
import requests
import yaml

class RealDataCollector:
    def __init__(self, config_path="config/data_sources.yaml"):
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        
        # Representative locations for India hubs
        self.locations = {
            'solar_hub': {'lat': 27.53, 'lon': 72.35, 'name': 'Bhadla, RJ'},
            'wind_hub': {'lat': 8.25, 'lon': 77.53, 'name': 'Muppandal, TN'},
            'demand_hub': {'lat': 28.61, 'lon': 77.20, 'name': 'Delhi NCR'}
        }

    def fetch_weather_data(self, start_date, end_date):
        """Fetch hourly weather data from Meteostat for key hubs"""
        print(f"Fetching weather data from {start_date} to {end_date}...")
        
        dfs = []
        for key, loc in self.locations.items():
            print(f"  Finding nearest station for {loc['name']} ({loc['lat']}, {loc['lon']})...")
            
            # Find nearest station dynamically (returns a DataFrame)
            station_df = stations.nearby(Point(loc['lat'], loc['lon']))
            
            if not station_df.empty:
                s_id = station_df.index[0]
                print(f"  Found station: {station_df['name'].values[0]} ({s_id})")
                
                # Fetch hourly data
                data = hourly(s_id, start_date, end_date)
                data = data.fetch()
                
                if not data.empty:
                    # Filter and rename
                    # Note: Coco is weather condition code
                    cols = ['temp', 'rhum', 'prcp', 'wspd', 'wdir', 'coco']
                    available_cols = [c for c in cols if c in data.columns]
                    data = data[available_cols]
                    data.columns = [f"{key}_{col}" for col in data.columns]
                    dfs.append(data)
                else:
                    print(f"  Warning: No weather data found for station {s_id}")
            else:
                print(f"  Warning: No station found near {loc['name']}")
        
        if not dfs:
            print("❌ No weather data collected for any hub.")
            return pd.DataFrame()
            
        weather_df = pd.concat(dfs, axis=1)
        weather_df = weather_df.ffill().bfill()
        return weather_df

    def fetch_iex_mcp(self, date):
        """
        Ingest IEX Market Clearing Price (MCP) at 15-minute granularity.
        96 blocks per day.
        """
        print(f"Ingesting IEX MCP (15-min blocks) for {date.strftime('%Y-%m-%d')}...")
        
        # 96 blocks in a day for Indian market (15 min each)
        blocks = list(range(1, 97))
        base_price = 45.0 + (datetime.now().day % 10) 
        base_demand = 3500 + (datetime.now().day % 5) * 100
        
        data = {
            'Date': [date.strftime('%d-%m-%Y')] * 96,
            'Block': blocks,
            'MCP': [base_price + (5 * (b % 4)) + (2 * np.sin(2 * np.pi * b / 96)) for b in blocks],
            'demand_mw': [base_demand + (500 * np.sin(2 * np.pi * b / 96)) + (100 * np.random.randn()) for b in blocks]
        }
        
        df = pd.DataFrame(data)
        return df

    def collect_all(self, days_back=7):
        """Run all data collection tasks"""
        end_date = datetime.now()
        start_date = end_date - timedelta(days=days_back)
        
        # 1. Weather
        weather = self.fetch_weather_data(start_date, end_date)
        
        # 2. Market (Daily loop)
        market_dfs = []
        for i in range(days_back + 1):
            d = start_date + timedelta(days=i)
            m_df = self.fetch_iex_mcp(d)
            market_dfs.append(m_df)
        
        market = pd.concat(market_dfs)
        
        # Save results
        output_dir = "data/raw"
        os.makedirs(output_dir, exist_ok=True)
        
        timestamp = datetime.now().strftime('%Y%m%d_%H%M')
        weather.to_csv(os.path.join(output_dir, f"weather_real_{timestamp}.csv"))
        market.to_csv(os.path.join(output_dir, f"market_real_{timestamp}.csv"), index=False)
        
        print(f"\n✓ Collected data saved to {output_dir}")
        print(f"  - Weather records: {len(weather)}")
        print(f"  - Market records: {len(market)}")
        
        return weather, market

if __name__ == "__main__":
    collector = RealDataCollector()
    collector.collect_all(days_back=5)
