requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
tensorflow>=2.13.0
torch>=2.0.0
statsmodels>=0.14.0
plotly>=5.14.0
dash>=2.11.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0
scipy>=1.11.0
optuna>=3.2.0
mlflow>=2.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
